Moving session mechanism (first steps)
