+++ core/web/server.py
@@ -1,7 +1,7 @@
 from socketserver import ThreadingMixIn
 from http.server import HTTPServer
-from core.web.handler import WebHandler
 from core.web import db
+from core.web.handler import WebHandler
 
 
 class ServerClass(ThreadingMixIn, HTTPServer):
@@ -13,13 +13,13 @@ class WebServer:
         self.logger = logger
         self.config = config
         self.HOST, self.PORT = config['host'], config['port']
+        db.register_client(config['db_address'])
         WebHandler.logger = self.logger
         WebHandler.SESS_EXP_TIME = config['auth_session_exp_time']
         self.server = ServerClass((self.HOST, self.PORT), WebHandler)
         self.server.allow_reuse_address = True
         self.server.version = config['version']
         self.server.www_dir = config['www_dir']
-        db.register_client(config['db_address'])
 
     def start(self):
         self.logger.info('Starting server on {}:{}'.format(self.server.server_name, self.server.server_port))
